requirements.txt
pandas
numpy
spacy
vaderSentiment
praw
pyyaml
scikit-learn
xgboost
lightgbm
backtrader
requests
tqdm
pytest 
tweepy